+++ app/Http/Controllers/IngredientController.php
@@ -76,7 +76,18 @@ class IngredientController extends Controller
     #[BAO\NotFoundResponse]
     public function show(Request $request, string $id): JsonResource
     {
-        $ingredient = Ingredient::with('cocktails', 'images', 'varieties', 'parentIngredient', 'createdUser', 'updatedUser', 'ingredientParts.ingredient', 'prices.priceCategory', 'category')
+        $ingredient = Ingredient::with(
+                'cocktails',
+                'images',
+                'varieties',
+                'parentIngredient',
+                'createdUser',
+                'updatedUser',
+                'ingredientParts.ingredient',
+                'prices.priceCategory',
+                'category',
+                'cocktailIngredientSubstitutes.cocktailIngredient.ingredient'
+            )
             ->withCount('cocktails')
             ->where('id', $id)
             ->orWhere('slug', $id)