+++ app/Http/Resources/IngredientResource.php
@@ -47,9 +47,16 @@ class IngredientResource extends JsonResource
                 'can_edit' => $request->user()->can('edit', $this->resource),
                 'can_delete' => $request->user()->can('delete', $this->resource),
             ]),
-            // 'in_shelf' => $this->when($this->relationLoaded('substitutes'), fn () => $this->userHasInShelf($request->user())),
             'in_shelf' => $this->userHasInShelf($request->user()),
             'in_shopping_list' => $this->userHasInShoppingList($request->user()),
+            'used_as_substitute_for' => $this->when(
+                $this->relationLoaded('cocktailIngredientSubstitutes'),
+                fn () => IngredientBasicResource::collection($this->getIngredientsUsedAsSubstituteFor())
+            ),
+            'can_be_substituted_with' => $this->when(
+                $this->relationLoaded('cocktailIngredientSubstitutes'),
+                fn () => IngredientBasicResource::collection($this->getCanBeSubstitutedWithIngredients())
+            ),
             'ingredient_parts' => $this->when(
                 $this->relationLoaded('ingredientParts'),
                 fn () => $this->ingredientParts->map(fn ($cip) => new IngredientBasicResource($cip->ingredient))