+++ CHANGELOG.md
@@ -78,8 +78,10 @@ This is a new major release. Here's a quick summary of the more interesting chan
     - Result is a JSON schema that can be imported via `/import/cocktail` endpoint
 - Added `/import/file` endpoint
     - This endpoint is used to import cocktails from a zip file containing files exported via `schema` type
+- Added `used_as_substitute_for` and `can_be_substituted_with` to ingredient response
 
 ## Changes
+- Recommended ingredients now take complex ingredients into consideration
 - Image processing moved from Imagick to Vips
     - This now requires `libvips` to be installed on the server
         - You can install it via `apt-get install -y --no-install-recommends libvips42`